Add NetworkAttachments to Octavia services
From other similar patches: For network isolation outgoing traffic,
service pods need to be attached additional networks. To achive this,
the the `k8s.v1.cni.cncf.io/networks` annotation needs to be set
on the pods.  This adds a networkAttachments parameter to the
CRD where it can be specified to which additional networks the
service pods should be attached to."

Note this also includes a fix to the amphoracontrollers for the Replicas
field.