Change the label of the octavia controllers
octavia-worker, healthmonitor and housekeeping services shared the same
label "octavia". But that was confusing as some resources like
NetworkAttachments are linked to the label and not the pod name. Each
service has now a unique label. That fixes a bug with the list of
NetworkAttachments as they all appeared only in the status of one pod.

The commit also fixed the json path for printing the NetworkAttachments