Fix a couple of router related issues.
Add missing negation of compare when reconciling router gateway info and
change the IP range of the provider subnet to avoid conflicts with IPs
allocated for pods and the macvlan master interfaces on the host. Also
fixes the physical netork name and changes the provider network and
routers to be created by the admin tenant.

Requires https://github.com/openstack-k8s-operators/install_yamls/pull/788