Send a gratuitous ARP when HM starts
As we are reusing the same ip address for the octavia
network-attachement when recreating the HM pod, we may have some issues
with the arp tables of the other resources on the network. Send a
gratuitous ARP when the octavia interface is ready, so the neutron
router is notified that the macaddr has changed.

Jira: OSPRH-5991