Predictable IPs for health managers and rsyslog containers
This patch implements predictable IPs for health managers by generating
a list of IPs on the same subnet as the octavia management network's
CIDR but outside of the existing IP ranges. These IPs mapped according
to the OpenShift node's name and the purpose (e.g. contro1_hm and
contro1_rsyslog). The IPs are set as aliases on the pods using the
downward API and a utility mounted into the pod.