Use my.cnf from config-data instead of subpath mount
Fixed a potential race condition in the octavia amphora controllers,
config-data is only mounted in the init container and my.cnf is mounted
as a subpath of the config-data mount in the "main" container. It may
trigger some errors because the config-data is unmounted when kubernetes
tries to mount the subpath.
Copy my.cnf from the config-data-merged mount, using the json config
file.

JIRA: OSPRH-9128