set securityContext.runAsNonRoot value to true
Reported by SAST tool:
Sigma main event: The Kubernetes container is allowed to run as the root
user. This may allow attackers to gain the root privileges of the host
when the container is compromised.
remediation: Explicitly set the 'securityContext.runAsNonRoot' value to
'true' to prevent the container from running as a root-level user.

JIRA: OSPRH-9907