Fixed order of MGMT_CIDR<n> environment variables
When setting the MGMT_CIDR<n> environment variable, we must ensure that
the entries are always defined in the same order, but
OctaviaProviderSubnetExtraCIDRs is built while iterating over a map, so
the order of the elements are not guaranteed.

When the order differs, the octavia-healthmanager pod is recreated
because its input parameters changed.

Sort the array that contains the data before setting the variable

JIRA: OSPRH-13951