Fix Client certificate duration
The client cert duration was only 1 month, it means that after one month
the octavia control plane could not communicate with the amphorae.
Extend this duration to 10 years (it was 5 years in Tripleo).
Add the client private key in the secret in case we need to update the
certificate without regenerating the CA.
Add a version tag in the secret.

JIRA: OSPRH-17062