Fix invalid flavorprofiles with empty image tag
octavia-operator set flavorprofile.amp_image_tag to an empty string when
a specific tag was not required, but octavia considers that the tag is
empty string instead of using the default tag, making the flavor
unusable.

Make amp_image_tag optional in its definition to not pass an empty
string when it's not set.

This commit also deals with updates, if octavia-operator detects that an
invalid flavorprofile exists, it deletes the associated flavor and
flavorprofile and recreates them correctly.

JIRA: OSPRH-17186